KUDU-3326 [delete]: Add soft delete table supports
Kudu system will not delete the table immediately after receiving the
command to delete the table. Instead, it will mark the table and set
a validity period. After the validity period, will try again to
determine whether the table really needs to be deleted.

NOTE: Soft-delete related functions is not supported when HMS is enabled.

Change-Id: I3d1dddfbca55a5c4bcac4028157325ad618ea665
Reviewed-on: http://gerrit.cloudera.org:8080/17917
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>