KUDU-3297 fix Thrift client used for HMS integration
As it turns out, in the context of KUDU-3297, the SASL negotiation code
needs to be updated in one more place:
  src/kudu/thrift/sasl_client_transport.cc

I also thought about unifying the code between the Thrift client
and the RPC code to have a single place to have the correct ordering
between the calls EnableProtection() and sasl_client_start(), but after
some consideration I realized it's not worth it.

As for the testing, I verified that before this patch the following
scenarios in hms_client-test were failing every time when running
on RedHat/CentOS 8.4:
  * ProtectionTypes/HmsClientTest.TestHmsOperations/1
  * ProtectionTypes/HmsClientTest.TestHmsOperations/3
  * ProtectionTypes/HmsClientTest.TestLargeObjects/1
  * ProtectionTypes/HmsClientTest.TestLargeObjects/3
The output of the failed test scenarios always contained the following:
  Bad status: Runtime error: failed to open Hive Metastore connection: SASL(-15): mechanism too weak for this user:

With this patch, all scenarios of the hms_client-test pass when running
on RedHat/CentOS 8.4:

This is a follow-up to fff48ea4e5eadd365a85a05a82f66b3eb76d0b0b.

Change-Id: Ic6af12932647eda7092f9f42a57eb211fe31f062
Reviewed-on: http://gerrit.cloudera.org:8080/17958
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>
(cherry picked from commit 0ade8c6f21f0887e90b261ae6b1a57f4a6d1eff1)
Reviewed-on: http://gerrit.cloudera.org:8080/18088
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>