[client] fix recently introduced ABI compatiblity issue
This patch fixes the recently introduced ABI compatibility issue in Kudu
C++ client once the KuduClient::ListTables() method was overloaded
(see [1] for details).  As I could see, the overloaded method was only
used in the Kudu CLI tooling with no apparent justification why it's
needed in the public client API.  With that, I moved the method to
be a part of the KuduClient::Data internal class.

In addition, I renamed --list_statistics into --show_table_info for the
`kudu table list` CLI tool and switched to the lower_camel_case for
the tags used to output a table's stats.

This is a follow-up to 3ccc434ea6dcbef15a12bd1734c8e66b69a611c5.

[1] https://community.kde.org/Policies/Binary_Compatibility_Issues_With_C%2B%2B

Change-Id: I214d424eb0e91b90db813fde0e0b7150253bca91
Reviewed-on: http://gerrit.cloudera.org:8080/17972
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>