[master] KUDU-3311 Allow to start with diff num of masters
Modified SysCatalogTable:Load to allow cluster startup if +1
master address is present in the flags.
The startup is halted, if the number of masters on-disk is higher than
what was provided in the flags, discouraging decommissioning masters
this way.

Additional manual test was done with the following steps in order to
ensure, that the documented way of removing a master works:
- ./start_kudu.sh -m 4 -t 3 -b $KUDU_BUILD_DIR -c /tmp/data --rpc-master 8800
- looked up the "last" master which was listening on :8006 and removed
  it by running:
  ./kudu master remove 127.0.0.1:8800,127.0.0.1:8802,127.0.0.1:8804,127.0.0.1:8806 127.0.0.1:8806
- stopped the cluster by running ./stop_kudu.sh
- restarted with 3 masters by running
  ./start_kudu.sh -m 3 -t 3 -b $KUDU_BUILD_DIR -c /tmp/data --rpc-master 8800

Change-Id: I39aeee2f52a55a8c29770f748895d38c9adff8a2
Reviewed-on: http://gerrit.cloudera.org:8080/17995
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins