KUDU-1959 - Add tests for /startup page and metrics for tservers
This patch implements the tests for the startup page using mini tablet
server.
 - We inject latency to bootstrap tablets while reading the webpage every
   10 milliseconds and validating the status for each step.
 - Fail a data directory and validate the status of each startup step.

We also validate the below startup metrics in the above scenarios
(log_block_manager* metrics in the case of using log block manager):
 - log_block_manager_total_containers_startup
 - log_block_manager_processed_containers_startup
 - log_block_manager_containers_processing_time_startup
 - tablets_num_total_startup
 - tablets_num_opened_startup
 - tablets_opening_time_startup

Additionally we also fix a race condition in the Kudu tablet server
WebUI. This race condition occurs if the tablet server is started while
the WebUI is continuously curled. The reason appears to be starting up
of the webserver before registering the path handlers as a part of
the change https://gerrit.cloudera.org/#/c/17730/

Change-Id: I9f432b4eb813e51214b4d6b3c5b7b4c89426f47f
Reviewed-on: http://gerrit.cloudera.org:8080/17990
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>