[tests] don't build client-stress-test with ASAN/TSAN
Test scenarios in client-stress-test are CPU and memory intensive,
and some of those are intended for benchmarking to track performance
regressions in the C++ client's metacache.  With that, there isn't much
sense to build and run them for ASAN/TSAN builds, so why to compile them
under ASAN/TSAN configurations at all?  As for the test coverage to catch
possible concurrency and undefined behavior issues, there is overlap
between the tests in client-stress-test and tests in client-test,
master-stress-test, and many others.

The motivation for this change is to address the feedback for one
of prior changelists [1].

  [1] https://gerrit.cloudera.org/#/c/18024/3/src/kudu/integration-tests/client-stress-test.cc@337

Change-Id: Ic044d051d6011d1cca95fc92d1daf400e3e3d116
Reviewed-on: http://gerrit.cloudera.org:8080/18027
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>