[client] fix custom hash schema criterion in KuduTableCreator::Create()
This patch fixes a bug in KuduTableCreator::Create(): the criterion
for the presence of custom hash schemas per range flag should account
for the case when the table-wide hash schema is non-empty, but there is
a range partition with empty hash schema.

This patch also contains a new test to cover the related functionality.
I verified that the test was failing before this patch, and with this
patch it's now passing.

Change-Id: I51fb745cfe7afb861b4bba54e56d79eca4832605
Reviewed-on: http://gerrit.cloudera.org:8080/18039
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>