[tests] harmonize assertions on the expected status
I was trying to understand why a test scenario from security-itest
failed in one of gerrit pre-commit runs and found that it would be
great to know the actual vs the expected type of the status when an
assertion of type ASSERT_{FALSE,TRUE}(s.IsXxx()) triggers.

This patch updates the corresponding call sites of ASSERT_{FALSE,TRUE}
in the code of that and other tests to provide such information for
easier troubleshooting and debugging in future.

Change-Id: I09cba638a5328fc57cc1ca70196bab42f09569b4
Reviewed-on: http://gerrit.cloudera.org:8080/18269
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins