KUDU-3328: make rebalancer tool take into account maintenance mode
At present kudu master would not assign new replicas on maintenance tservers
but we could still move replicas to them through the rebalancer tool.
Sometimes we want this kind of operation to be disabled because the maintenance
tservers would be restarted or decommissioned in a short time.

This patch improves the rebalancer tool's behavior. If maintenance tservers are
not specified in 'ignored_tservers' and users don't specify
'-force_rebalance_replicas_on_maintenance_tservers' neither, the rebalancer tool
couldn't get replica moves and return IllegalState.

Change-Id: I843a2521a811ab67c5f98bb43cc84367edb3fe0b
Reviewed-on: http://gerrit.cloudera.org:8080/18074
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>