[consensus] reduce critical section in LogCache::ReadOps()
This patch reduces the section protected by the lock in
LogCache::ReadOps(), moving out of the critical section several
operations which do not need such a protection.

I didn't find an existing performance test which would exercise the
updated code path, but reducing the time spent under a lock looks like
an improvement.

Change-Id: Ia7de32eb8adde242c8047720e062535cbdaf3e39
Reviewed-on: http://gerrit.cloudera.org:8080/18090
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins