[tools] Kudu table schema in Avro format
This patch introduces a new flag -avro_format_schema to the existing
"kudu table describe" CLI tool to output the table schema in Avro
format. The other information displayed by the tool like owner,
comments, partitions won't be displayed if this flag is set to true.
This flag cannot be used in conjunction with other flags.

Change-Id: I98194739e8132f93e916d1a78d338d28dd9e8075
Reviewed-on: http://gerrit.cloudera.org:8080/18121
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>