[tool] Add --host to start_kudu.sh script
The start_kudu.sh script currently works with an assumption that the
Kudu servers will be accessed from the same host that they're running on
(i.e. localhost).

Recently, while running Kudu in a remote Linux box, I wanted to access
it from my Mac, but the ports were bound to the loopback interface, so I
wasn't able to do that. This change makes it possible to change what
interface the servers bound to. To make sure the web server can be
connected to from both a remote machine and from localhost, this patch
also removes the --webserver_interface flag, which defaults to 0.0.0.0.

Change-Id: I68f920a03f494d277b5f7008b0a75cd927634d9f
Reviewed-on: http://gerrit.cloudera.org:8080/18127
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>