[util] add scheduling priority metric for a server
This patch introduces a new metric: the scheduling priority for a Kudu
server process.  The new metric's key is "scheduling_priority" for the
JSON output, and the value is what getpriority(PRIO_PROCESS, getpid())
returns.  With the new metric it's possible to tell whether Kudu masters
and tablet servers are being run with altered scheduling priority
(e.g., using the nice/renice tools).

Change-Id: I985511636a9099779a811eabe6785cf765ec04b7
Reviewed-on: http://gerrit.cloudera.org:8080/18144
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>