[clang-tidy] add modernize-{deprecated-headers,use-override}
This patch adds modernize-deprecated-headers [1] and
modernize-use-override [2] checks for clang-tidy.

The former helps with changing the C-style headers into C++ ones: that
complements the IWYU functionality since the currently used IWYU version
suggests adding C-style headers, not C++ ones, and it's not clear what's
the timeline for addressing that.

The latter helps to remove extra 'virtual' keywords: I saw that in
a few recent patches posted for review, so I thought it would be nice
to automate catching those nits.

[1] https://releases.llvm.org/9.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/modernize-deprecated-headers.html
[2] https://releases.llvm.org/9.0.0/tools/clang/tools/extra/docs/clang-tidy/checks/modernize-use-override.html

Change-Id: I04e8e2e19b185350969a2945326a7193d1e8f60d
Reviewed-on: http://gerrit.cloudera.org:8080/18182
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>