[schema] use operator==() instead of Equals()
While looking at the code in src/kudu/common/schema.h in the scope of
reviewing a prior changelist, I noticed that the comparison options
for Schema::Equals() were used only in the test scenarios.  I took
the liberty and moved the Equals() method into the test code and
replaced it with the operator== at call sites.  This patch includes
other minor cleanup, but overall the patch doesn't contain any
functional changes.

Change-Id: Ib0e3d75d216241efec49b64233198eafe52871cb
Reviewed-on: http://gerrit.cloudera.org:8080/18237
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>