[tools] correct rebalancer's initial status message
I noticed that the 'kudu cluster rebalance' tool was reporting that
it had been '...running whole-cluster rebalancing...' even when running
with a particular set of tables to rebalance, i.e. not a whole-cluster
rebalancing by any means if there is more than one table in a cluster.

This patch fixes the issue.

Change-Id: Ic224a07f9ce9ed3ca9a447022f183b799a1bebee
Reviewed-on: http://gerrit.cloudera.org:8080/18297
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>