[tools] add beginning of the key range into TabletSummary
This patch adds the beginning of the key range into the TabletSummary
class used in ksck reporting.  Corresponding TabletSummaryPB has been
updated as well.

This patch also contains a test to verify the presence of the newly
added field in TabletSummary entries produced by ksck.

The newly added field will be used in a follow-up changelist for
range-aware rebalancing.

Change-Id: Icf083e1ae33919200f6e0b0f0a667ee6eea1bab3
Reviewed-on: http://gerrit.cloudera.org:8080/18356
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>