[cmake_modules] add '-o /dev/null' to the linker command
I noticed that every time I run cmake in the build directory
on macOS, a.out file appears.  It turned out it's generated
by the GET_LINKER_VERSION procedure from KuduLinker.cmake.

This patch add '-o /dev/null' to the command that determines
the version of the linker to avoid generating a.out file
in the current working directory.

Change-Id: Id095902679a2f0e2648c45bcd4024fc734b9a1e7
Reviewed-on: http://gerrit.cloudera.org:8080/18394
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>