Add .asf.yaml
An .asf.yaml was committed to the kudu-site repo from which the site
is published some time ago, but this is deleted when regenerating the
site from kudu/gh-pages. Adding this file in kudu/gh-pages should
resolve this issue.

This file is a config file required by the ASF infra[1], without it,
the updated site wouldn't be published.

[1] https://cwiki.apache.org/confluence/display/INFRA/Git+-+.asf.yaml+features#Git.asf.yamlfeatures-WebsitedeploymentserviceforGitrepositories

Change-Id: I58d8f85e83ac2ad46d192abb58b748eb7bba6c98
Reviewed-on: http://gerrit.cloudera.org:8080/18410
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Attila Bukor <abukor@apache.org>