[master] KUDU-2671: Range specific hashing during table alter op.
This commit has the proto file, master.proto, updated to reflect
the changes needed for having custom hash partitioning while
adding a new range to an existing table.
The AlterTableRequestPB has the new field to accommodate the
new change.
The reason for pushing this patch with just this change is to
unblock client side work while the server side work is being
done.

Change-Id: Ifec5566ea8f3e49d00dcb6964b3d17c4be0504eb
Reviewed-on: http://gerrit.cloudera.org:8080/18485
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Kudu Jenkins