[docs] update a few notes on multi-master migration
This patch updates the warning note for the multi-master migration
section (of course, that's for 1.15 and newer versions) to bring more
clarity on what components to be restarted if not using DNS aliases.
The prior '... bringing the cluster down for maintenance ...' was not
technically correct since it's possible to perform rolling restart of
tablet servers in the cluster.

In addition, the obsolete warning note about the multi-master
migration procedure being unsafe is removed.  The new multi-master
migration procedure (for 1.15 and newer versions) is universal and
safe when migrating from N to N+1 masters.

Change-Id: Ie707e8342adf7a5e557079a4121a50d009f7602f
Reviewed-on: http://gerrit.cloudera.org:8080/18487
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>