KUDU-3367 [compaction] add supplement to gc algorithm
If we get a REDO delta full of delete ops, which means there is
not a single update operation in the delta. The current compaction
algorithm doesn't run GC on such deltamemstores. The accumulation
of deltamemstores like that negatively affects performance of scan
operations.

This patch as a supplement to KUDU-1625, we could  release storage
space for old tablet metadata that does not support the live count
function. See KUDU-3367 for details.

Change-Id: I8b26737dffecc17688b42188da959b2ba16351ed
Reviewed-on: http://gerrit.cloudera.org:8080/18503
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>