KUDU-2671 relax requirement for partition range specification
As of now (see [1]), the number of hash dimensions in a per-range custom
hash schema must be the same for all the ranges in a table.  With that,
it's possible to relax the requirements on specifying partition key
ranges for a GetTableLocations RPC.  That allows legacy Kudu clients
to work with tables having custom hash schemas per range.  If necessary,
it's possible to make the requirements stricter by setting
--require_new_spec_for_custom_hash_schema_range_bound=true.

This is a follow-up to [1].

[1] https://github.com/apache/kudu/commit/6998193e69eeda497f912d1d806470c95b591ad4

Change-Id: Idfdae92f6baee5e8f197cba41c8e51120f9b4d58
Reviewed-on: http://gerrit.cloudera.org:8080/18509
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>