KUDU-3368 Encrypt file keys with server keys
This patch removes the dummy encryption key used to encrypt all file
keys and introduces the server key, which is used instead on a
per-server basis. The server key is stored in the instance files,
currently in cleartext. Encrypting the server key will be done in a
follow-up patch.

Servers in internal mini-cluster are forced to use the same server key
as they share an environment, but servers in external mini-cluster can
use different keys. Multiple (non-default) PosixEnv instances can now be
created to make this possible.

Change-Id: I1884f62fde3bb110291b1d01c5e68942c6071318
Reviewed-on: http://gerrit.cloudera.org:8080/18192
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>